MAINT: Moved the EthereumAddress trait back to utils_t as @moria pointed out
