+++ enigma-crypto/src/hash.rs
@@ -35,20 +35,6 @@ pub trait Sha256<T> {
     fn sha256(&self) -> T where T: Sized;
 }
 
-pub trait EthereumAddress<T> {
-    fn address(&self) -> T where T: Sized;
-}
-
-
-impl EthereumAddress<String> for [u8; 64] {
-    // TODO: Maybe add a checksum address
-    fn address(&self) -> String {
-        let mut result: String = String::from("0x");
-        result.push_str(&self.keccak256()[12..32].to_hex::<String>());
-        result
-    }
-}
-
 impl Keccak256<[u8; 32]> for [u8] {
     fn keccak256(&self) -> [u8; 32] {
         let mut keccak = Keccak::new_keccak256();
